Rename Client to WebRTCClient
In preparation for exploring if RTSPtoWebRTC is better
